requirements.txt
mcp==1.9.0
pandas==2.2.3
pymongo==4.12.1
PyMySQL==1.1.1
psycopg2==2.9.10
gspread==6.1.4
gspread-dataframe==4.0.0
google-cloud-bigquery==3.33.0
db-dtypes==1.4.2
